more user profile design and refactoring

+++ bonobo/bonobo/Views/UserInfoPage.xaml.cs
@@ -15,16 +15,24 @@ namespace bonobo.Views
         public UserInfoPage()
         {
             InitializeComponent();
-            Init();
+            UserProfileView user = new UserProfileView
+            {
+                FirstName = App.UserDatabase.GetUser().FirstName,
+                LastName = App.UserDatabase.GetUser().LastName,
+                Gender = App.UserDatabase.GetUser().Gender,
+                Birthdate = App.UserDatabase.GetUser().BirthDate,
+                HeaderImage = "http://acephalous.typepad.com/.a/6a00d8341c2df453ef017d3c2bd399970c-500wi",
+                ProfileImage = "https://manofmany.com/wp-content/uploads/2017/07/Jon-Snow-2.jpg",
+                Reviews = 16,
+                Hosted = 11,
+                Joined = 24
+            };
+
+            Init(user);
         }
 
-        void Init()
+        void Init(UserProfileView user)
         {
-            UserProfileView user = new UserProfileView(
-                "Jon", 
-                "Snow", 
-                "http://acephalous.typepad.com/.a/6a00d8341c2df453ef017d3c2bd399970c-500wi", "https://manofmany.com/wp-content/uploads/2017/07/Jon-Snow-2.jpg");
-
             Img_ProfilePicture.TranslationY = 100;
             Img_ProfilePicture.HeightRequest = 100;
             Img_ProfilePicture.WidthRequest = 100;
@@ -37,13 +45,27 @@ namespace bonobo.Views
             Img_ProfileHeader.Aspect = Aspect.AspectFill;
             Img_ProfileHeader.HeightRequest = 250;
 
-            Lbl_FirstName.Text = user.FirstName;
-            Lbl_LastName.Text = user.LastName;
+            Lbl_Name.Text = user.FirstName + " " + user.LastName;
+            Lbl_Name.FontAttributes = FontAttributes.Bold;
+
             Lbl_TagLine.Text = "I conquer super villains and make the world a safer place.";
-            Lbl_Email.Text = App.UserDatabase.GetUser().FirstName
-                + " "
-                + App.UserDatabase.GetUser().LastName
-                + App.UserDatabase.GetUser().BirthDate.ToString();
+            Lbl_TagLine.FontAttributes = FontAttributes.Italic;
+
+            Lbl_Reviews.Text = user.Reviews.ToString();
+            Lbl_Reviews.TextColor = Constants.MainTextColor;
+            Lbl_Joined.Text = user.Joined.ToString();
+            Lbl_Joined.TextColor = Constants.MainTextColor;
+            Lbl_Hosted.Text =user.Hosted.ToString();
+            Lbl_Hosted.TextColor = Constants.MainTextColor;
+
+            Lbl_Desc.Text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse malesuada ultricies arcu nec egestas. Nam porta fermentum aliquam. Nullam tincidunt odio purus.";
+
+            //compute user age
+            DateTime zeroTime = new DateTime(1, 1, 1);
+            TimeSpan span = DateTime.Now - user.Birthdate;
+            int years = (zeroTime + span).Year - 1;
+            Lbl_Birthday.Text = "Age: " + years.ToString();
+            Lbl_Gender.Text = "Gender: " + user.Gender;
         }
 
         async void OnTapGestureForSigningOut(object sender, EventArgs args)
@@ -64,7 +86,7 @@ namespace bonobo.Views
                 await DisplayAlert("Logout", "Not able to reach server in time.", "Ok");
             }
 
-            //if (result)
+            if (result)
             {
                 //delete logged user from local storage
                 App.UserDatabase.DeleteUser(0);
